requirements.txt
APScheduler==3.10.1
miniupnpc==2.0.2
python-dotenv==1.0.0
requests==2.28.2
